Cleanupe les utils
